requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.0.0
click>=8.1.0
pyyaml>=6.0
cryptography>=41.0.0
bcrypt>=4.0.0
pyotp>=2.9.0
qrcode[pil]>=7.4.0
apscheduler>=3.10.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
itsdangerous>=2.1.0
starlette>=0.27.0

